backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg[binary]>=3.1.0
python-decouple==3.8
django-cors-headers==4.3.1
django-filter==23.5
Pillow>=10.0.0
feedparser==6.0.11
beautifulsoup4==4.12.3
httpx==0.27.0
tweepy==4.14.0
anthropic>=0.39.0
